+++ lib/src/client/compatible_client.dart
@@ -1,6 +1,7 @@
 import 'package:http/http.dart';
 import 'package:rhttp/src/client/rhttp_client.dart';
 import 'package:rhttp/src/interceptor/interceptor.dart';
+import 'package:rhttp/src/model/cancel_token.dart';
 import 'package:rhttp/src/model/exception.dart';
 import 'package:rhttp/src/model/request.dart';
 import 'package:rhttp/src/model/settings.dart';
@@ -20,7 +21,9 @@ class RhttpCompatibleClient with BaseClient {
   /// The actual client that is used to make requests.
   final RhttpClient client;
 
-  RhttpCompatibleClient._(this.client);
+  /// Creates a new HTTP client from an existing [RhttpClient].
+  /// Only for advanced use cases.
+  RhttpCompatibleClient.of(this.client);
 
   /// Creates a new HTTP client asynchronously.
   /// Use this method if your app is already running to avoid blocking the UI.
@@ -32,7 +35,7 @@ class RhttpCompatibleClient with BaseClient {
       settings: (settings ?? const ClientSettings()).digest(),
       interceptors: interceptors,
     );
-    return RhttpCompatibleClient._(client);
+    return RhttpCompatibleClient.of(client);
   }
 
   /// Creates a new HTTP client synchronously.
@@ -41,7 +44,7 @@ class RhttpCompatibleClient with BaseClient {
   ///
   /// Note:
   /// This method crashes when configured to use HTTP/3.
-  /// See: https://github.com/Tienisto/rhttp/issues/10
+  /// See: https://codeberg.org/Tienisto/rhttp/issues/10
   factory RhttpCompatibleClient.createSync({
     ClientSettings? settings,
     List<Interceptor>? interceptors,
@@ -50,23 +53,56 @@ class RhttpCompatibleClient with BaseClient {
       settings: (settings ?? const ClientSettings()).digest(),
       interceptors: interceptors,
     );
-    return RhttpCompatibleClient._(client);
+    return RhttpCompatibleClient.of(client);
   }
 
   @override
   Future<StreamedResponse> send(BaseRequest request) async {
+    final CancelToken? cancelToken;
+    switch (request) {
+      case Abortable(abortTrigger: final trigger?):
+        cancelToken = CancelToken();
+        trigger.then((_) {
+          cancelToken?.cancel();
+        });
+        break;
+      case _:
+        cancelToken = null;
+    }
+
     try {
       final response = await client.requestStream(
         method: HttpMethod(request.method.toUpperCase()),
         url: request.url.toString(),
         headers: HttpHeaders.rawMap(request.headers),
         body: HttpBody.bytes(await request.finalize().toBytes()),
+        cancelToken: cancelToken,
       );
 
-      final responseHeaderMap = response.headerMap;
+      final responseHeaderMap = response.headerMapList.map(
+        // Combine multiple values as per `http` package specification.
+        (key, value) => MapEntry(key, value.join(', ')),
+      );
 
       return StreamedResponse(
-        response.body,
+        response.body.handleError((e, st) {
+          if (e is RhttpException) {
+            if (e is RhttpCancelException) {
+              Error.throwWithStackTrace(
+                RhttpWrappedRequestAbortedException(request.url, e),
+                st,
+              );
+            }
+            Error.throwWithStackTrace(
+              RhttpWrappedClientException(e.toString(), request.url, e),
+              st,
+            );
+          }
+          Error.throwWithStackTrace(
+            ClientException(e.toString(), request.url),
+            st,
+          );
+        }),
         response.statusCode,
         contentLength: switch (responseHeaderMap['content-length']) {
           String s => int.parse(s),
@@ -83,6 +119,12 @@ class RhttpCompatibleClient with BaseClient {
         reasonPhrase: null,
       );
     } on RhttpException catch (e, st) {
+      if (e is RhttpCancelException) {
+        Error.throwWithStackTrace(
+          RhttpWrappedRequestAbortedException(request.url, e),
+          st,
+        );
+      }
       Error.throwWithStackTrace(
         RhttpWrappedClientException(e.toString(), request.url, e),
         st,
@@ -108,6 +150,17 @@ class RhttpWrappedClientException extends ClientException {
   final RhttpException rhttpException;
 
   RhttpWrappedClientException(super.message, super.uri, this.rhttpException);
+
+  @override
+  String toString() => rhttpException.toString();
+}
+
+/// Special case for [RequestAbortedException].
+class RhttpWrappedRequestAbortedException extends RequestAbortedException {
+  /// The original exception that was thrown by rhttp.
+  final RhttpException rhttpException;
+
+  RhttpWrappedRequestAbortedException(super.uri, this.rhttpException);
 
   @override
   String toString() => rhttpException.toString();